requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptocompare==0.7.6
cryptography==41.0.1
cycler==0.11.0
dateparser==1.1.8
Django==4.2.2
fonttools==4.40.0
frozendict==2.3.8
frozenlist==1.3.3
future==0.18.3
gTTS==2.3.2
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
line-bot-sdk==2.4.2
lxml==4.9.2
matplotlib==3.7.1
mplfinance==0.12.9b7
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.3
openai==0.28.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.18.0
pydub==0.25.1
pyimgur==0.6.0
pyngrok==6.0.0
pyparsing==3.0.9
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
sqlparse==0.4.4
tiktoken==0.4.0
tqdm==4.66.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.26.16
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.2
yfinance==0.2.20
